+++ backend/src/index.js
@@ -1,12 +1,20 @@
-const express = require('express');
+import express from 'express';
 
 const app = express();
-const PORT = process.env.PORT || 3000;
+app.use(express.json());
 
-app.get('/', (req, res) => {
-  res.send('Hello, world!');
+const workPeriods = [
+  {
+    id: 1, startTime: '2024-08-10T08:00:00Z', endTime: '2024-08-10T16:00:00Z', jobId: 101, userId: 1,
+  },
+  {
+    id: 2, startTime: '2024-08-11T09:00:00Z', endTime: '2024-08-11T17:00:00Z', jobId: 102, userId: 2,
+  },
+];
+
+// Define routes in index.js
+app.get('/work-periods', async (req, res) => {
+  res.json(workPeriods);
 });
 
-app.listen(PORT, () => {
-  console.log('Server is running on port ${PORT}');
-});
+export default app;  // Export the app instance
